Separated out normalization asts into their own file
