+++ crates/generate_artifacts/src/generate_artifacts.rs
@@ -51,6 +51,9 @@ lazy_static! {
     pub static ref ENTRYPOINT: ArtifactFilePrefix = "entrypoint".intern().into();
     pub static ref ISO_TS_FILE_NAME: ArtifactFileName = "iso.ts".intern().into();
     pub static ref ISO_TS: ArtifactFilePrefix = "iso".intern().into();
+    pub static ref NORMALIZATION_AST_FILE_NAME: ArtifactFileName =
+        "normalization_ast.ts".intern().into();
+    pub static ref NORMALIZATION_AST: ArtifactFilePrefix = "normalization_ast".intern().into();
     pub static ref QUERY_TEXT_FILE_NAME: ArtifactFileName = "query_text.ts".intern().into();
     pub static ref QUERY_TEXT: ArtifactFilePrefix = "query_text".intern().into();
     pub static ref REFETCH_READER_FILE_NAME: ArtifactFileName = "refetch_reader.ts".intern().into();
