+++ demos/pet-demo/src/components/__isograph/AdItem/AdItemDisplay/entrypoint.ts
@@ -3,52 +3,9 @@ import {AdItem__AdItemDisplay__param} from './param_type';
 import {AdItem__AdItemDisplay__output_type} from './output_type';
 import readerResolver from './resolver_reader';
 import queryText from './query_text';
+import normalizationAst from './normalization_ast';
 const nestedRefetchQueries: RefetchQueryNormalizationArtifactWrapper[] = [];
 
-const normalizationAst: NormalizationAst = {
-  kind: "NormalizationAst",
-  selections: [
-    {
-      kind: "Linked",
-      fieldName: "node",
-      arguments: [
-        [
-          "id",
-          { kind: "Variable", name: "id" },
-        ],
-      ],
-      concreteType: null,
-      selections: [
-        {
-          kind: "InlineFragment",
-          type: "AdItem",
-          selections: [
-            {
-              kind: "Scalar",
-              fieldName: "__typename",
-              arguments: null,
-            },
-            {
-              kind: "Scalar",
-              fieldName: "id",
-              arguments: null,
-            },
-            {
-              kind: "Scalar",
-              fieldName: "advertiser",
-              arguments: null,
-            },
-            {
-              kind: "Scalar",
-              fieldName: "message",
-              arguments: null,
-            },
-          ],
-        },
-      ],
-    },
-  ],
-};
 const artifact: IsographEntrypoint<
   AdItem__AdItemDisplay__param,
   AdItem__AdItemDisplay__output_type,
