+++ demos/pet-demo/src/components/__isograph/Mutation/SetTagline/entrypoint.ts
@@ -3,44 +3,9 @@ import {Mutation__SetTagline__param} from './param_type';
 import {Mutation__SetTagline__output_type} from './output_type';
 import readerResolver from './resolver_reader';
 import queryText from './query_text';
+import normalizationAst from './normalization_ast';
 const nestedRefetchQueries: RefetchQueryNormalizationArtifactWrapper[] = [];
 
-const normalizationAst: NormalizationAst = {
-  kind: "NormalizationAst",
-  selections: [
-    {
-      kind: "Linked",
-      fieldName: "set_pet_tagline",
-      arguments: [
-        [
-          "input",
-          { kind: "Variable", name: "input" },
-        ],
-      ],
-      concreteType: "SetPetTaglineResponse",
-      selections: [
-        {
-          kind: "Linked",
-          fieldName: "pet",
-          arguments: null,
-          concreteType: "Pet",
-          selections: [
-            {
-              kind: "Scalar",
-              fieldName: "id",
-              arguments: null,
-            },
-            {
-              kind: "Scalar",
-              fieldName: "tagline",
-              arguments: null,
-            },
-          ],
-        },
-      ],
-    },
-  ],
-};
 const artifact: IsographEntrypoint<
   Mutation__SetTagline__param,
   Mutation__SetTagline__output_type,
