+++ demos/pet-demo/src/components/__isograph/Mutation/SetTagline/normalization_ast.ts
@@ -1,0 +1,38 @@
+import type {NormalizationAst} from '@isograph/react';
+const normalizationAst: NormalizationAst = {
+  kind: "NormalizationAst",
+  selections: [
+    {
+      kind: "Linked",
+      fieldName: "set_pet_tagline",
+      arguments: [
+        [
+          "input",
+          { kind: "Variable", name: "input" },
+        ],
+      ],
+      concreteType: "SetPetTaglineResponse",
+      selections: [
+        {
+          kind: "Linked",
+          fieldName: "pet",
+          arguments: null,
+          concreteType: "Pet",
+          selections: [
+            {
+              kind: "Scalar",
+              fieldName: "id",
+              arguments: null,
+            },
+            {
+              kind: "Scalar",
+              fieldName: "tagline",
+              arguments: null,
+            },
+          ],
+        },
+      ],
+    },
+  ],
+};
+export default normalizationAst;
