+++ demos/pet-demo/src/components/__isograph/Pet/PetCheckinsCardList/entrypoint.ts
@@ -3,78 +3,12 @@ import {Pet__PetCheckinsCardList__param} from './param_type';
 import {Pet__PetCheckinsCardList__output_type} from './output_type';
 import readerResolver from './resolver_reader';
 import queryText from './query_text';
+import normalizationAst from './normalization_ast';
 import refetchQuery0 from './__refetch__0';
 const nestedRefetchQueries: RefetchQueryNormalizationArtifactWrapper[] = [
   { artifact: refetchQuery0, allowedVariables: ["checkin_id", ] },
 ];
 
-const normalizationAst: NormalizationAst = {
-  kind: "NormalizationAst",
-  selections: [
-    {
-      kind: "Linked",
-      fieldName: "node",
-      arguments: [
-        [
-          "id",
-          { kind: "Variable", name: "id" },
-        ],
-      ],
-      concreteType: null,
-      selections: [
-        {
-          kind: "InlineFragment",
-          type: "Pet",
-          selections: [
-            {
-              kind: "Scalar",
-              fieldName: "__typename",
-              arguments: null,
-            },
-            {
-              kind: "Scalar",
-              fieldName: "id",
-              arguments: null,
-            },
-            {
-              kind: "Linked",
-              fieldName: "checkins",
-              arguments: [
-                [
-                  "skip",
-                  { kind: "Variable", name: "skip" },
-                ],
-
-                [
-                  "limit",
-                  { kind: "Variable", name: "limit" },
-                ],
-              ],
-              concreteType: "Checkin",
-              selections: [
-                {
-                  kind: "Scalar",
-                  fieldName: "id",
-                  arguments: null,
-                },
-                {
-                  kind: "Scalar",
-                  fieldName: "location",
-                  arguments: null,
-                },
-                {
-                  kind: "Scalar",
-                  fieldName: "time",
-                  arguments: null,
-                },
-              ],
-            },
-          ],
-        },
-      ],
-    },
-  ],
-};
 const artifact: IsographEntrypoint<
   Pet__PetCheckinsCardList__param,
   Pet__PetCheckinsCardList__output_type,
