+++ demos/pet-demo/src/components/__isograph/Query/HomeRoute/entrypoint.ts
@@ -3,41 +3,9 @@ import {Query__HomeRoute__param} from './param_type';
 import {Query__HomeRoute__output_type} from './output_type';
 import readerResolver from './resolver_reader';
 import queryText from './query_text';
+import normalizationAst from './normalization_ast';
 const nestedRefetchQueries: RefetchQueryNormalizationArtifactWrapper[] = [];
 
-const normalizationAst: NormalizationAst = {
-  kind: "NormalizationAst",
-  selections: [
-    {
-      kind: "Linked",
-      fieldName: "pets",
-      arguments: null,
-      concreteType: "Pet",
-      selections: [
-        {
-          kind: "Scalar",
-          fieldName: "id",
-          arguments: null,
-        },
-        {
-          kind: "Scalar",
-          fieldName: "name",
-          arguments: null,
-        },
-        {
-          kind: "Scalar",
-          fieldName: "picture",
-          arguments: null,
-        },
-        {
-          kind: "Scalar",
-          fieldName: "tagline",
-          arguments: null,
-        },
-      ],
-    },
-  ],
-};
 const artifact: IsographEntrypoint<
   Query__HomeRoute__param,
   Query__HomeRoute__output_type,
