+++ demos/pet-demo/src/components/__isograph/Query/HomeRoute/normalization_ast.ts
@@ -1,0 +1,35 @@
+import type {NormalizationAst} from '@isograph/react';
+const normalizationAst: NormalizationAst = {
+  kind: "NormalizationAst",
+  selections: [
+    {
+      kind: "Linked",
+      fieldName: "pets",
+      arguments: null,
+      concreteType: "Pet",
+      selections: [
+        {
+          kind: "Scalar",
+          fieldName: "id",
+          arguments: null,
+        },
+        {
+          kind: "Scalar",
+          fieldName: "name",
+          arguments: null,
+        },
+        {
+          kind: "Scalar",
+          fieldName: "picture",
+          arguments: null,
+        },
+        {
+          kind: "Scalar",
+          fieldName: "tagline",
+          arguments: null,
+        },
+      ],
+    },
+  ],
+};
+export default normalizationAst;
