+++ demos/pet-demo/src/components/__isograph/Query/Newsfeed/entrypoint.ts
@@ -3,109 +3,9 @@ import {Query__Newsfeed__param} from './param_type';
 import {Query__Newsfeed__output_type} from './output_type';
 import readerResolver from './resolver_reader';
 import queryText from './query_text';
+import normalizationAst from './normalization_ast';
 const nestedRefetchQueries: RefetchQueryNormalizationArtifactWrapper[] = [];
 
-const normalizationAst: NormalizationAst = {
-  kind: "NormalizationAst",
-  selections: [
-    {
-      kind: "Linked",
-      fieldName: "viewer",
-      arguments: null,
-      concreteType: "Viewer",
-      selections: [
-        {
-          kind: "Scalar",
-          fieldName: "id",
-          arguments: null,
-        },
-        {
-          kind: "Linked",
-          fieldName: "newsfeed",
-          arguments: [
-            [
-              "skip",
-              { kind: "Literal", value: 0 },
-            ],
-
-            [
-              "limit",
-              { kind: "Literal", value: 6 },
-            ],
-          ],
-          concreteType: null,
-          selections: [
-            {
-              kind: "Scalar",
-              fieldName: "__typename",
-              arguments: null,
-            },
-            {
-              kind: "InlineFragment",
-              type: "AdItem",
-              selections: [
-                {
-                  kind: "Scalar",
-                  fieldName: "id",
-                  arguments: null,
-                },
-                {
-                  kind: "Scalar",
-                  fieldName: "__typename",
-                  arguments: null,
-                },
-              ],
-            },
-            {
-              kind: "InlineFragment",
-              type: "BlogItem",
-              selections: [
-                {
-                  kind: "Scalar",
-                  fieldName: "id",
-                  arguments: null,
-                },
-                {
-                  kind: "Scalar",
-                  fieldName: "__typename",
-                  arguments: null,
-                },
-                {
-                  kind: "Scalar",
-                  fieldName: "author",
-                  arguments: null,
-                },
-                {
-                  kind: "Scalar",
-                  fieldName: "content",
-                  arguments: null,
-                },
-                {
-                  kind: "Linked",
-                  fieldName: "image",
-                  arguments: null,
-                  concreteType: "Image",
-                  selections: [
-                    {
-                      kind: "Scalar",
-                      fieldName: "id",
-                      arguments: null,
-                    },
-                  ],
-                },
-                {
-                  kind: "Scalar",
-                  fieldName: "title",
-                  arguments: null,
-                },
-              ],
-            },
-          ],
-        },
-      ],
-    },
-  ],
-};
 const artifact: IsographEntrypoint<
   Query__Newsfeed__param,
   Query__Newsfeed__output_type,
