+++ demos/pet-demo/src/components/__isograph/Query/Newsfeed/normalization_ast.ts
@@ -1,0 +1,103 @@
+import type {NormalizationAst} from '@isograph/react';
+const normalizationAst: NormalizationAst = {
+  kind: "NormalizationAst",
+  selections: [
+    {
+      kind: "Linked",
+      fieldName: "viewer",
+      arguments: null,
+      concreteType: "Viewer",
+      selections: [
+        {
+          kind: "Scalar",
+          fieldName: "id",
+          arguments: null,
+        },
+        {
+          kind: "Linked",
+          fieldName: "newsfeed",
+          arguments: [
+            [
+              "skip",
+              { kind: "Literal", value: 0 },
+            ],
+
+            [
+              "limit",
+              { kind: "Literal", value: 6 },
+            ],
+          ],
+          concreteType: null,
+          selections: [
+            {
+              kind: "Scalar",
+              fieldName: "__typename",
+              arguments: null,
+            },
+            {
+              kind: "InlineFragment",
+              type: "AdItem",
+              selections: [
+                {
+                  kind: "Scalar",
+                  fieldName: "id",
+                  arguments: null,
+                },
+                {
+                  kind: "Scalar",
+                  fieldName: "__typename",
+                  arguments: null,
+                },
+              ],
+            },
+            {
+              kind: "InlineFragment",
+              type: "BlogItem",
+              selections: [
+                {
+                  kind: "Scalar",
+                  fieldName: "id",
+                  arguments: null,
+                },
+                {
+                  kind: "Scalar",
+                  fieldName: "__typename",
+                  arguments: null,
+                },
+                {
+                  kind: "Scalar",
+                  fieldName: "author",
+                  arguments: null,
+                },
+                {
+                  kind: "Scalar",
+                  fieldName: "content",
+                  arguments: null,
+                },
+                {
+                  kind: "Linked",
+                  fieldName: "image",
+                  arguments: null,
+                  concreteType: "Image",
+                  selections: [
+                    {
+                      kind: "Scalar",
+                      fieldName: "id",
+                      arguments: null,
+                    },
+                  ],
+                },
+                {
+                  kind: "Scalar",
+                  fieldName: "title",
+                  arguments: null,
+                },
+              ],
+            },
+          ],
+        },
+      ],
+    },
+  ],
+};
+export default normalizationAst;
