+++ demos/pet-demo/src/components/__isograph/Query/PetByName/entrypoint.ts
@@ -3,36 +3,9 @@ import {Query__PetByName__param} from './param_type';
 import {Query__PetByName__output_type} from './output_type';
 import readerResolver from './resolver_reader';
 import queryText from './query_text';
+import normalizationAst from './normalization_ast';
 const nestedRefetchQueries: RefetchQueryNormalizationArtifactWrapper[] = [];
 
-const normalizationAst: NormalizationAst = {
-  kind: "NormalizationAst",
-  selections: [
-    {
-      kind: "Linked",
-      fieldName: "petByName",
-      arguments: [
-        [
-          "name",
-          { kind: "Variable", name: "name" },
-        ],
-      ],
-      concreteType: "Pet",
-      selections: [
-        {
-          kind: "Scalar",
-          fieldName: "id",
-          arguments: null,
-        },
-        {
-          kind: "Scalar",
-          fieldName: "name",
-          arguments: null,
-        },
-      ],
-    },
-  ],
-};
 const artifact: IsographEntrypoint<
   Query__PetByName__param,
   Query__PetByName__output_type,
