+++ demos/pet-demo/src/components/__isograph/Query/PetByName/normalization_ast.ts
@@ -1,0 +1,30 @@
+import type {NormalizationAst} from '@isograph/react';
+const normalizationAst: NormalizationAst = {
+  kind: "NormalizationAst",
+  selections: [
+    {
+      kind: "Linked",
+      fieldName: "petByName",
+      arguments: [
+        [
+          "name",
+          { kind: "Variable", name: "name" },
+        ],
+      ],
+      concreteType: "Pet",
+      selections: [
+        {
+          kind: "Scalar",
+          fieldName: "id",
+          arguments: null,
+        },
+        {
+          kind: "Scalar",
+          fieldName: "name",
+          arguments: null,
+        },
+      ],
+    },
+  ],
+};
+export default normalizationAst;
