+++ demos/pet-demo/src/components/__isograph/Query/PetCheckinListRoute/entrypoint.ts
@@ -3,67 +3,12 @@ import {Query__PetCheckinListRoute__param} from './param_type';
 import {Query__PetCheckinListRoute__output_type} from './output_type';
 import readerResolver from './resolver_reader';
 import queryText from './query_text';
+import normalizationAst from './normalization_ast';
 import refetchQuery0 from './__refetch__0';
 const nestedRefetchQueries: RefetchQueryNormalizationArtifactWrapper[] = [
   { artifact: refetchQuery0, allowedVariables: ["checkin_id", ] },
 ];
 
-const normalizationAst: NormalizationAst = {
-  kind: "NormalizationAst",
-  selections: [
-    {
-      kind: "Linked",
-      fieldName: "pet",
-      arguments: [
-        [
-          "id",
-          { kind: "Variable", name: "id" },
-        ],
-      ],
-      concreteType: "Pet",
-      selections: [
-        {
-          kind: "Scalar",
-          fieldName: "id",
-          arguments: null,
-        },
-        {
-          kind: "Linked",
-          fieldName: "checkins",
-          arguments: [
-            [
-              "skip",
-              { kind: "Literal", value: 0 },
-            ],
-
-            [
-              "limit",
-              { kind: "Literal", value: 1 },
-            ],
-          ],
-          concreteType: "Checkin",
-          selections: [
-            {
-              kind: "Scalar",
-              fieldName: "id",
-              arguments: null,
-            },
-            {
-              kind: "Scalar",
-              fieldName: "location",
-              arguments: null,
-            },
-          ],
-        },
-        {
-          kind: "Scalar",
-          fieldName: "name",
-          arguments: null,
-        },
-      ],
-    },
-  ],
-};
 const artifact: IsographEntrypoint<
   Query__PetCheckinListRoute__param,
   Query__PetCheckinListRoute__output_type,
