+++ demos/pet-demo/src/components/__isograph/Query/PetCheckinListRoute/normalization_ast.ts
@@ -1,0 +1,58 @@
+import type {NormalizationAst} from '@isograph/react';
+const normalizationAst: NormalizationAst = {
+  kind: "NormalizationAst",
+  selections: [
+    {
+      kind: "Linked",
+      fieldName: "pet",
+      arguments: [
+        [
+          "id",
+          { kind: "Variable", name: "id" },
+        ],
+      ],
+      concreteType: "Pet",
+      selections: [
+        {
+          kind: "Scalar",
+          fieldName: "id",
+          arguments: null,
+        },
+        {
+          kind: "Linked",
+          fieldName: "checkins",
+          arguments: [
+            [
+              "skip",
+              { kind: "Literal", value: 0 },
+            ],
+
+            [
+              "limit",
+              { kind: "Literal", value: 1 },
+            ],
+          ],
+          concreteType: "Checkin",
+          selections: [
+            {
+              kind: "Scalar",
+              fieldName: "id",
+              arguments: null,
+            },
+            {
+              kind: "Scalar",
+              fieldName: "location",
+              arguments: null,
+            },
+          ],
+        },
+        {
+          kind: "Scalar",
+          fieldName: "name",
+          arguments: null,
+        },
+      ],
+    },
+  ],
+};
+export default normalizationAst;
