+++ demos/pet-demo/src/components/__isograph/Query/PetDetailDeferredRoute/entrypoint.ts
@@ -3,63 +3,9 @@ import {Query__PetDetailDeferredRoute__param} from './param_type';
 import {Query__PetDetailDeferredRoute__output_type} from './output_type';
 import readerResolver from './resolver_reader';
 import queryText from './query_text';
+import normalizationAst from './normalization_ast';
 const nestedRefetchQueries: RefetchQueryNormalizationArtifactWrapper[] = [];
 
-const normalizationAst: NormalizationAst = {
-  kind: "NormalizationAst",
-  selections: [
-    {
-      kind: "Linked",
-      fieldName: "pet",
-      arguments: [
-        [
-          "id",
-          { kind: "Variable", name: "id" },
-        ],
-      ],
-      concreteType: "Pet",
-      selections: [
-        {
-          kind: "Scalar",
-          fieldName: "id",
-          arguments: null,
-        },
-        {
-          kind: "Scalar",
-          fieldName: "name",
-          arguments: null,
-        },
-      ],
-    },
-    {
-      kind: "Linked",
-      fieldName: "topLevelField",
-      arguments: [
-        [
-          "input",
-          {
-            kind: "Object",
-            value: [
-              [
-                "name",
-                { kind: "String", value: "ThisIsJustHereToTestObjectLiterals" },
-              ],
-
-            ]
-          },
-        ],
-      ],
-      concreteType: "TopLevelField",
-      selections: [
-        {
-          kind: "Scalar",
-          fieldName: "__typename",
-          arguments: null,
-        },
-      ],
-    },
-  ],
-};
 const artifact: IsographEntrypoint<
   Query__PetDetailDeferredRoute__param,
   Query__PetDetailDeferredRoute__output_type,
