+++ demos/pet-demo/src/components/__isograph/Query/PetDetailRoute/entrypoint.ts
@@ -3,6 +3,7 @@ import {Query__PetDetailRoute__param} from './param_type';
 import {Query__PetDetailRoute__output_type} from './output_type';
 import readerResolver from './resolver_reader';
 import queryText from './query_text';
+import normalizationAst from './normalization_ast';
 import refetchQuery0 from './__refetch__0';
 import refetchQuery1 from './__refetch__1';
 import refetchQuery2 from './__refetch__2';
@@ -16,179 +17,6 @@ const nestedRefetchQueries: RefetchQueryNormalizationArtifactWrapper[] = [
   { artifact: refetchQuery4, allowedVariables: ["id", ] },
 ];
 
-const normalizationAst: NormalizationAst = {
-  kind: "NormalizationAst",
-  selections: [
-    {
-      kind: "Linked",
-      fieldName: "pet",
-      arguments: [
-        [
-          "id",
-          { kind: "Variable", name: "id" },
-        ],
-      ],
-      concreteType: "Pet",
-      selections: [
-        {
-          kind: "Scalar",
-          fieldName: "id",
-          arguments: null,
-        },
-        {
-          kind: "Scalar",
-          fieldName: "age",
-          arguments: null,
-        },
-        {
-          kind: "Linked",
-          fieldName: "best_friend_relationship",
-          arguments: null,
-          concreteType: "BestFriendRelationship",
-          selections: [
-            {
-              kind: "Linked",
-              fieldName: "best_friend",
-              arguments: null,
-              concreteType: "Pet",
-              selections: [
-                {
-                  kind: "Scalar",
-                  fieldName: "id",
-                  arguments: null,
-                },
-                {
-                  kind: "Scalar",
-                  fieldName: "name",
-                  arguments: null,
-                },
-                {
-                  kind: "Scalar",
-                  fieldName: "picture",
-                  arguments: null,
-                },
-              ],
-            },
-            {
-              kind: "Scalar",
-              fieldName: "picture_together",
-              arguments: null,
-            },
-          ],
-        },
-        {
-          kind: "Linked",
-          fieldName: "checkins",
-          arguments: [
-            [
-              "skip",
-              { kind: "Literal", value: null },
-            ],
-
-            [
-              "limit",
-              { kind: "Literal", value: null },
-            ],
-          ],
-          concreteType: "Checkin",
-          selections: [
-            {
-              kind: "Scalar",
-              fieldName: "id",
-              arguments: null,
-            },
-            {
-              kind: "Scalar",
-              fieldName: "location",
-              arguments: null,
-            },
-            {
-              kind: "Scalar",
-              fieldName: "time",
-              arguments: null,
-            },
-          ],
-        },
-        {
-          kind: "Scalar",
-          fieldName: "favorite_phrase",
-          arguments: null,
-        },
-        {
-          kind: "Scalar",
-          fieldName: "name",
-          arguments: null,
-        },
-        {
-          kind: "Scalar",
-          fieldName: "nickname",
-          arguments: null,
-        },
-        {
-          kind: "Linked",
-          fieldName: "potential_new_best_friends",
-          arguments: null,
-          concreteType: "Pet",
-          selections: [
-            {
-              kind: "Scalar",
-              fieldName: "id",
-              arguments: null,
-            },
-            {
-              kind: "Scalar",
-              fieldName: "name",
-              arguments: null,
-            },
-          ],
-        },
-        {
-          kind: "Linked",
-          fieldName: "stats",
-          arguments: null,
-          concreteType: "PetStats",
-          selections: [
-            {
-              kind: "Scalar",
-              fieldName: "cuteness",
-              arguments: null,
-            },
-            {
-              kind: "Scalar",
-              fieldName: "energy",
-              arguments: null,
-            },
-            {
-              kind: "Scalar",
-              fieldName: "hunger",
-              arguments: null,
-            },
-            {
-              kind: "Scalar",
-              fieldName: "intelligence",
-              arguments: null,
-            },
-            {
-              kind: "Scalar",
-              fieldName: "sociability",
-              arguments: null,
-            },
-            {
-              kind: "Scalar",
-              fieldName: "weight",
-              arguments: null,
-            },
-          ],
-        },
-        {
-          kind: "Scalar",
-          fieldName: "tagline",
-          arguments: null,
-        },
-      ],
-    },
-  ],
-};
 const artifact: IsographEntrypoint<
   Query__PetDetailRoute__param,
   Query__PetDetailRoute__output_type,
