+++ libs/isograph-react/src/tests/__isograph/Query/subquery/entrypoint.ts
@@ -3,44 +3,9 @@ import {Query__subquery__param} from './param_type';
 import {Query__subquery__output_type} from './output_type';
 import readerResolver from './resolver_reader';
 import queryText from './query_text';
+import normalizationAst from './normalization_ast';
 const nestedRefetchQueries: RefetchQueryNormalizationArtifactWrapper[] = [];
 
-const normalizationAst: NormalizationAst = {
-  kind: "NormalizationAst",
-  selections: [
-    {
-      kind: "Linked",
-      fieldName: "query",
-      arguments: null,
-      concreteType: "Query",
-      selections: [
-        {
-          kind: "Linked",
-          fieldName: "node",
-          arguments: [
-            [
-              "id",
-              { kind: "Variable", name: "id" },
-            ],
-          ],
-          concreteType: null,
-          selections: [
-            {
-              kind: "Scalar",
-              fieldName: "__typename",
-              arguments: null,
-            },
-            {
-              kind: "Scalar",
-              fieldName: "id",
-              arguments: null,
-            },
-          ],
-        },
-      ],
-    },
-  ],
-};
 const artifact: IsographEntrypoint<
   Query__subquery__param,
   Query__subquery__output_type,
