+++ libs/isograph-react/src/tests/__isograph/Query/subquery/normalization_ast.ts
@@ -1,0 +1,38 @@
+import type {NormalizationAst} from '@isograph/react';
+const normalizationAst: NormalizationAst = {
+  kind: "NormalizationAst",
+  selections: [
+    {
+      kind: "Linked",
+      fieldName: "query",
+      arguments: null,
+      concreteType: "Query",
+      selections: [
+        {
+          kind: "Linked",
+          fieldName: "node",
+          arguments: [
+            [
+              "id",
+              { kind: "Variable", name: "id" },
+            ],
+          ],
+          concreteType: null,
+          selections: [
+            {
+              kind: "Scalar",
+              fieldName: "__typename",
+              arguments: null,
+            },
+            {
+              kind: "Scalar",
+              fieldName: "id",
+              arguments: null,
+            },
+          ],
+        },
+      ],
+    },
+  ],
+};
+export default normalizationAst;
